Use typed function for `jwt.VerifyOption`
This is partially aesthetic, but I believe it's better for IDEs to read when used in a function signature, which will  hopefully make it easier to understand for any users of this library.

Using a type alias makes it less obvious and reading the function signature, it's not as obvious what the arguments are used for and what functional options a user should attempt to use.